+++ src/util/delay.rs
@@ -4,19 +4,28 @@ use std::{
     task::{Context, Poll},
     time::Duration,
 };
+use pin_project_lite::pin_project;
 
-/// A runtime-agnostic delay.
-///
-/// This helps abstract over different runtime implementations and only uses
-/// [`async_io::Timer`] as a fallback implementation. [`async_io::Timer`] is
-/// generally avoided, since most runtimes provide their own timer
-/// implementations without needing to spawn a new helper thread.
-pub(crate) struct Delay(DelayInner);
+pin_project! {
+    /// A runtime-agnostic delay.
+    ///
+    /// This helps abstract over different runtime implementations and only uses
+    /// [`async_io::Timer`] as a fallback implementation. [`async_io::Timer`] is
+    /// generally avoided, since most runtimes provide their own timer
+    /// implementations without needing to spawn a new helper thread.
+    pub(crate) struct Delay {
+        #[pin]
+        inner: DelayInner
+    }
+}
 
-enum DelayInner {
-    #[cfg(feature = "tokio-timer")]
-    Tokio(Pin<Box<tokio::time::Sleep>>),
-    AsyncIo(Pin<Box<async_io::Timer>>),
+pin_project_cfg! {
+    #[project = DelayInnerProj]
+    pub(crate) enum DelayInner {
+        #[cfg(feature = "tokio-timer")]
+        Tokio { #[pin] fut: tokio::time::Sleep, },
+        AsyncIo { #[pin] fut: async_io::Timer, },
+    }
 }
 
 impl Delay {
@@ -24,25 +33,28 @@ impl Delay {
     pub fn new(duration: Duration) -> Self {
         #[cfg(feature = "tokio-timer")]
         if tokio::runtime::Handle::try_current().is_ok() {
-            return Self(DelayInner::Tokio(Box::pin(tokio::time::sleep(duration))));
+            return Self { inner: DelayInner::Tokio {
+                fut: tokio::time::sleep(duration),
+            }};
         }
 
         // Fallback
         #[allow(unreachable_code)]
-        Self(DelayInner::AsyncIo(Box::pin(async_io::Timer::after(
-            duration,
-        ))))
+        Self { inner: DelayInner::AsyncIo {
+            fut: async_io::Timer::after(duration),
+        }}
     }
 
     /// Resets the delay to wait for the specified duration.
-    pub fn reset(&mut self, duration: Duration) {
-        match &mut self.0 {
+    pub fn reset(self: Pin<&mut Self>, duration: Duration) {
+        let this = self.project();
+        match this.inner.project() {
             #[cfg(feature = "tokio-timer")]
-            DelayInner::Tokio(sleep) => {
-                sleep.as_mut().reset(tokio::time::Instant::now() + duration);
+            DelayInnerProj::Tokio { fut: sleep } => {
+                sleep.reset(tokio::time::Instant::now() + duration);
             }
-            DelayInner::AsyncIo(timer) => {
-                timer.as_mut().set_after(duration);
+            DelayInnerProj::AsyncIo { fut: mut timer } => {
+                timer.set_after(duration);
             }
         }
     }
@@ -51,14 +63,12 @@ impl Delay {
 impl Future for Delay {
     type Output = ();
 
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        match &mut self.0 {
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let this = self.project();
+        match this.inner.project() {
             #[cfg(feature = "tokio-timer")]
-            DelayInner::Tokio(sleep) => sleep.as_mut().poll(cx),
-            DelayInner::AsyncIo(timer) => timer.as_mut().poll(cx).map(|_| ()),
+            DelayInnerProj::Tokio { fut: sleep } => sleep.poll(cx),
+            DelayInnerProj::AsyncIo { fut: timer } => timer.poll(cx).map(|_| ()),
         }
     }
 }
-
-// Ensure Delay is Unpin
-impl Unpin for Delay {}